Simple SMP judge
Judge using 2 cores. One for ducksrv, another for contestant.
Use shared memory to communicate, and use IPI to trap the process.
Must run without hyper-threading enabled.

TODOs: IPI with more functions, hyper-threading detection, syscalls through shared memory.

+++ kern/pmap.c
@@ -402,7 +402,7 @@ page_init(void)
 	
 	assert(MPENTRY_PADDR % PGSIZE == 0);
 	for(int i = npages - 1; i >= 0; i--)
-		if(i == 0 || (i * PGSIZE >= IOPHYSMEM && i * PGSIZE < EXTPHYSMEM + (64 << 20)) || i * PGSIZE == MPENTRY_PADDR)
+		if(i == 0 || ((uint32_t) i * PGSIZE >= IOPHYSMEM && (uint32_t) i * PGSIZE < EXTPHYSMEM + (64 << 20)) || i * PGSIZE == MPENTRY_PADDR)
 		{
 			pages[i].pp_ref = 233; // 233?
 		}
